requirements.txt
pytest-playwright
playwright-stealth
stem
PySocks
bs4
